src/api/requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
pydantic
python-dotenv
requests
